move CloningRepository out to it's own model
