+++ app/src/models/cloning-repository.ts
@@ -1,0 +1,28 @@
+import * as Path from 'path'
+
+let CloningRepositoryID = 1
+
+/** A repository which is currently being cloned. */
+export class CloningRepository {
+  public readonly id = CloningRepositoryID++
+  public readonly path: string
+  public readonly url: string
+
+  public constructor(path: string, url: string) {
+    this.path = path
+    this.url = url
+  }
+
+  public get name(): string {
+    return Path.basename(this.path)
+  }
+
+  /**
+   * A hash of the properties of the object.
+   *
+   * Objects with the same hash are guaranteed to be structurally equal.
+   */
+  public get hash(): string {
+    return `${this.id}+${this.path}+${this.url}`
+  }
+}
