+++ app/src/ui/app.tsx
@@ -5,7 +5,7 @@ import { RepositoriesList } from './repositories-list'
 import { RepositoryView } from './repository'
 import { TitleBar } from './window/title-bar'
 import { Dispatcher } from '../lib/dispatcher'
-import { AppStore, CloningRepository } from '../lib/stores'
+import { AppStore } from '../lib/stores'
 import { Repository } from '../models/repository'
 import { MenuEvent } from '../main-process/menu'
 import { assertNever } from '../lib/fatal-error'
@@ -44,9 +44,11 @@ import { UpdateAvailable } from './updates'
 import { Preferences } from './preferences'
 import { Account } from '../models/account'
 import { TipState } from '../models/tip'
+import { CloningRepository } from '../models/cloning-repository'
 import { shouldRenderApplicationMenu } from './lib/features'
 import { Merge } from './merge-branch'
 import { RepositorySettings } from './repository-settings'
+import { matchExistingRepository } from '../lib/repository-matching'
 import { AppError } from './app-error'
 import { MissingRepository } from './missing-repository'
 import { AddExistingRepository, CreateRepository } from './add-repository'
